% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_similarity.R
\name{get_text_similarity}
\alias{get_text_similarity}
\title{Calculate how similar each user's text messeges are to each other.}
\usage{
get_text_similarity(user_ids, content, quiet = FALSE)
}
\arguments{
\item{user_ids}{A \code{vector} of type \code{character}.}

\item{content}{A \code{vector} of type \code{character} containing the text
content of the messages.}

\item{quiet}{If \code{FALSE}, prints status information.}
}
\value{
A \code{tibble} containing user ids, number of tweets per user and
similarity distance.
}
\description{
Calculate how similar each user's text messeges are to each other.
}
