% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_indicators.R
\name{get_tweets_per_day}
\alias{get_tweets_per_day}
\title{Calculate how similar each user's text messeges are to each other.}
\usage{
get_tweets_per_day(user_ids, times, quiet = FALSE)
}
\arguments{
\item{user_ids}{A \code{vector} of type \code{character}.}

\item{times}{A \code{vector} of type \code{character} containing the time
stamps of the messages.}

\item{quiet}{If \code{FALSE}, prints status information.}
}
\value{
A \code{tibble} containing a user id and the number of tweets per day
for each user.
}
\description{
Calculate how similar each user's text messeges are to each other.
}
